requirements.txt
requests
dnspython
pymysql
DBUtils
celery
flower
flask
gevent
pyyaml